Refactor 3 classes and added into diffrent folder
