+++ associations/book.rb
@@ -7,7 +7,8 @@ class Book
     @rentals = []
   end
 
-  def add_rental(person, date)
-    @rentals.push(Rental.new(date, self, person))
+  def add_rental(rental)
+    @rentals << rental
+    rental.book = self
   end
 end
