--- classroom.rb
@@ -1,13 +1,0 @@
-class Classroom
-  attr_accessor :label, :students
-
-  def initialize(label)
-    @label = label
-    @students = []
-  end
-
-  def add_student(student)
-    student.classroom = self
-    @students.push(student)
-  end
-end
