--- rental.rb
@@ -1,11 +1,0 @@
-class Rental
-  attr_accessor :date, :book, :person
-
-  def initialize(date, book, person)
-    @date = date
-    @book = book
-    @person = person
-    @book.rentals.push(self)
-    @person.rentals.push(self)
-  end
-end
